pipelines/requirements.txt
aiohttp==3.9.1
boto3==1.33.13
requests==2.31.0
Pillow>=9.0.0

# Development dependencies
